Change rollout strategy to Recreate
Jellyfin is a stateful application that can't handle multiple instances
accessing the same database file. The default rollout strategy brings up
a new pod before killing the old one. This creates a
situation where the new pod is stuck in a CrashLoopBackoff because it
can't acquire a lock on the database.

This *does* mean that rollouts will always have some downtime, but that
was the reality of the situation anyways. Once Jellyfin supports
multiple instances then this could be changed back to the default value.